Extract OrganizationEntry into its own component
