+++ src/components/mod.rs
@@ -1,2 +1,3 @@
 pub mod welcome;
 pub mod about;
+pub mod organization_entry;
